requirements.txt
numpy<2.0.0
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.30.0
pyannote.audio==3.3.2
hf_transfer
requests
ffmpeg-python
fastapi
uvicorn
pydantic
prometheus-client
prometheus-fastapi-instrumentator
tenacity
python-multipart
aiofiles
python-dotenv
yt-dlp
modal
browser_cookie3